% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcharts-package.R
\docType{package}
\name{funcharts-package}
\alias{funcharts}
\alias{funcharts-package}
\title{funcharts: Functional Control Charts}
\description{
Provides functional control charts for statistical process monitoring of functional data, using the methods of Capezza et al. (2020) \doi{10.1002/asmb.2507}, Centofanti et al. (2021) \doi{10.1080/00401706.2020.1753581}, Capezza et al. (2024) \doi{10.1080/00224065.2024.2383674}, Capezza et al. (2024) \doi{10.1080/00401706.2024.2327346}, Centofanti et al. (2025) \doi{10.1080/00224065.2024.2430978}, Capezza et al. (2025) \doi{10.48550/arXiv.2410.20138}. The package is thoroughly illustrated in the paper of Capezza et al (2023) \doi{10.1080/00224065.2023.2219012}.
}
\references{
Capezza C, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2020)
Control charts for
monitoring ship operating conditions and CO2 emissions based
on scalar-on-function regression.
\emph{Applied Stochastic Models in Business and Industry},
36(3):477--500.
\url{doi:10.1002/asmb.2507}

Centofanti F, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2021)
Functional Regression Control Chart.
\emph{Technometrics}, 63(3):281--294, \url{doi:10.1080/00401706.2020.1753581}

Capezza, C., Centofanti, F., Lepore, A., Palumbo, B. (2024)
Robust Multivariate Functional Control Chart.
\emph{Technometrics}, 66(4):531--547, \url{doi:10.1080/00401706.2024.2327346}.

Capezza, C., Capizzi, G., Centofanti, F., Lepore, A., Palumbo, B. (2025)
An Adaptive Multivariate Functional EWMA Control Chart.
\emph{Journal of Quality Technology},  57(1):1--15,
\url{doi:10.1080/00224065.2024.2383674}.

Centofanti, F., A. Lepore, M. Kulahci, and M. P. Spooner (2024).
Real-time monitoring of functional data. \emph{Journal of Quality Technology}, 57(2):135--152,
\url{doi:10.1080/00224065.2024.2430978}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/unina-sfere/funcharts}
  \item Report bugs at \url{https://github.com/unina-sfere/funcharts/issues}
}

}
\author{
\strong{Maintainer}: Christian Capezza \email{christian.capezza@unina.it}

Authors:
\itemize{
  \item Fabio Centofanti
  \item Davide Forcina
  \item Antonio Lepore
  \item Biagio Palumbo
}

Other contributors:
\itemize{
  \item Alessandra Menafoglio [contributor]
  \item Simone Vantini [contributor]
}

}
\keyword{internal}
