% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcharts.R
\docType{package}
\name{funcharts}
\alias{funcharts}
\title{\code{funcharts} package}
\description{
Provides functional control charts
for statistical process monitoring of functional data,
using the methods of Capezza et al. (2020) \url{doi:10.1002/asmb.2507} and
Centofanti et al. (2021) \url{doi:10.1080/00401706.2020.1753581}.
The package is thoroughly illustrated in the paper of
Capezza et al (2023) \url{doi:10.1080/00224065.2023.2219012}
}
\references{
Capezza C, Centofanti F, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2023)
funcharts: control charts for multivariate functional data in R.
\emph{Journal of Quality Technology}, \url{doi:10.1002/asmb.2507}

Capezza C, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2020)
Control charts for
monitoring ship operating conditions and CO2 emissions based
on scalar-on-function regression.
\emph{Applied Stochastic Models in Business and Industry},
36(3):477--500.
\url{doi:10.1002/asmb.2507}

Centofanti F, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2021)
Functional Regression Control Chart.
\emph{Technometrics}, 63(3), 281--294. \url{doi:10.1080/00401706.2020.1753581}
}
