% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{get_mfd_fd}
\alias{get_mfd_fd}
\title{Convert a \code{fd} object into a Multivariate Functional Data object.}
\usage{
get_mfd_fd(fdobj)
}
\arguments{
\item{fdobj}{An object of class fd.}
}
\value{
An object of class \code{mfd}.
See also \code{?mfd} for additional details on the
multivariate functional data class.
}
\description{
Convert a \code{fd} object into a Multivariate Functional Data object.
}
\examples{
library(funcharts)
library(fda)
bs <- create.bspline.basis(nbasis = 10)
fdobj <- fd(coef = 1:10, basisobj = bs)
mfdobj <- get_mfd_fd(fdobj)
}
\seealso{
\code{mfd}
}
