% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_fault_detection.R
\name{get_ooc}
\alias{get_ooc}
\title{Get out of control observations from control charts}
\usage{
get_ooc(cclist)
}
\arguments{
\item{cclist}{A \code{data.frame} produced by
\code{\link{control_charts_pca}}, \code{\link{control_charts_sof_pc}},
\code{\link{regr_cc_fof}}, or \code{\link{regr_cc_sof}}.}
}
\value{
A \code{data.frame} with the same number of rows as cclist,
and the same number of columns
apart from the columns indicating control chart limits.
Each value is TRUE if the corresponding observation is in control
and FALSE otherwise.
}
\description{
Get out of control observations from control charts
}
\examples{
library(funcharts)
data("air")
air <- lapply(air, function(x) x[1:220, , drop = FALSE])
fun_covariates <- c("CO", "temperature")
mfdobj_x <- get_mfd_list(air[fun_covariates],
                         n_basis = 15,
                         lambda = 1e-2)
mfdobj_y <- get_mfd_list(air["NO2"],
                         n_basis = 15,
                         lambda = 1e-2)
y_scalar <- rowMeans(air$NO2)
y1_scalar <- y_scalar[1:200]
y2_scalar <- y_scalar[201:220]
mfdobj_y1 <- mfdobj_y[1:200]
mfdobj_y2 <- mfdobj_y[201:220]
mfdobj_x1 <- mfdobj_x[1:200]
mfdobj_x2 <- mfdobj_x[201:220]
mod_fof <- fof_pc(mfdobj_y1, mfdobj_x1)
cclist <- regr_cc_fof(mod_fof,
                      mfdobj_y_new = mfdobj_y2,
                      mfdobj_x_new = mfdobj_x2)
get_ooc(cclist)


}
