% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{is.mfd}
\alias{is.mfd}
\title{Confirm Object has Class \code{mfd}}
\usage{
is.mfd(mfdobj)
}
\arguments{
\item{mfdobj}{An object to be checked.}
}
\value{
a logical value: TRUE if the class is correct, FALSE otherwise.
}
\description{
Check that an argument is a multivariate
functional data object of class \code{mfd}.
}
