% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{lines.mfd}
\alias{lines.mfd}
\title{Add curves to an existing multivariate functional data plot}
\usage{
\method{lines}{mfd}(x, ...)
}
\arguments{
\item{x}{An \code{mfd} object.}

\item{...}{Graphical arguments passed to \code{\link{plot.mfd}} with
\code{add=TRUE}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Adds new curves from an \code{mfd} object to an existing plot created by
\code{\link{plot.mfd}}. Each variable is added to its corresponding panel,
using the same scales and layout as the original plot.
}
\seealso{
\code{\link{plot.mfd}}, \code{\link{abline_mfd}}
}
