% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{lines_mfd}
\alias{lines_mfd}
\title{Add the plot of a new multivariate functional data object to an existing
plot.}
\usage{
lines_mfd(
  plot_mfd_obj,
  mfdobj_new,
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = TRUE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  type_mfd = "mfd",
  y_lim_equal = FALSE,
  ...
)
}
\arguments{
\item{plot_mfd_obj}{A plot produced by \code{link{plot_mfd}}}

\item{mfdobj_new}{A new multivariate functional data object of class mfd to be plotted.}

\item{mapping}{See \code{\link{plot_mfd}}.}

\item{data}{See \code{\link{plot_mfd}}.}

\item{stat}{See \code{\link{plot_mfd}}.}

\item{position}{See \code{\link{plot_mfd}}.}

\item{na.rm}{See \code{\link{plot_mfd}}.}

\item{orientation}{See \code{\link{plot_mfd}}.}

\item{show.legend}{See \code{\link{plot_mfd}}.}

\item{inherit.aes}{See \code{\link{plot_mfd}}.}

\item{type_mfd}{See \code{\link{plot_mfd}}.}

\item{y_lim_equal}{See \code{\link{plot_mfd}}.}

\item{...}{See \code{\link{plot_mfd}}.}
}
\value{
A plot of the multivariate functional data object added to the existing
one.
}
\description{
Add the plot of a new multivariate functional data object to an existing
plot.
}
\examples{
library(funcharts)
library(ggplot2)
mfdobj1 <- data_sim_mfd()
mfdobj2 <- data_sim_mfd()
p <- plot_mfd(mfdobj1)
lines_mfd(p, mfdobj_new = mfdobj2)

}
