% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{nobs.mfd}
\alias{nobs.mfd}
\title{Number of observations in a multivariate functional data object}
\usage{
\method{nobs}{mfd}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mfd}.}

\item{...}{Further arguments passed to methods (not used).}
}
\value{
An integer: the number of observations.
}
\description{
Number of observations in a multivariate functional data object
}
\examples{
# nobs(mfdobj)
}
