% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{norm.mfd}
\alias{norm.mfd}
\title{Norm of Multivariate Functional Data}
\usage{
norm.mfd(mfdobj)
}
\arguments{
\item{mfdobj}{A multivariate functional data object of class \code{mfd}.}
}
\value{
A vector of length equal to the number of replications
in \code{mfdobj},
containing the norm of each multivariate functional observation
in the product Hilbert space,
i.e. the sum of \emph{L^2} norms for each functional variable.
}
\description{
Norm of multivariate functional data contained
in a \code{mfd} object.
}
\examples{
library(funcharts)
mfdobj <- data_sim_mfd()
norm.mfd(mfdobj)

}
