% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_fault_detection.R
\name{which_ooc}
\alias{which_ooc}
\title{Get the index of the out of control observations from control charts}
\usage{
which_ooc(cclist)
}
\arguments{
\item{cclist}{A \code{data.frame} produced by
\code{\link{control_charts_sof_pc}}.}
}
\value{
A list of as many \code{data.frame} objects as
the control charts in \code{cclist}.
Each data frame has two columns, the \code{n} contains
an index number giving the observation in the
phase II data set, i.e. 1 for the first observation,
2 for the second, and so on,
while the \code{id} column contains the id of the observation,
which can be general and
depends on the specific data set.
}
\description{
This function returns a list for each control chart and returns
the id of all observations that are out of control in that control chart.
}
\examples{
library(funcharts)
data("air")
air <- lapply(air, function(x) x[1:220, , drop = FALSE])
fun_covariates <- c("CO", "temperature")
mfdobj_x <- get_mfd_list(air[fun_covariates],
                         n_basis = 15,
                         lambda = 1e-2)
mfdobj_y <- get_mfd_list(air["NO2"],
                         n_basis = 15,
                         lambda = 1e-2)
y_scalar <- rowMeans(air$NO2)
y1_scalar <- y_scalar[1:200]
y2_scalar <- y_scalar[201:220]
mfdobj_y1 <- mfdobj_y[1:200]
mfdobj_y2 <- mfdobj_y[201:220]
mfdobj_x1 <- mfdobj_x[1:200]
mfdobj_x2 <- mfdobj_x[201:220]
mod_fof <- fof_pc(mfdobj_y1, mfdobj_x1)
cclist <- regr_cc_fof(mod_fof,
                      mfdobj_y_new = mfdobj_y2,
                      mfdobj_x_new = mfdobj_x2)
which_ooc(cclist)

}
