% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcharts-package.R
\docType{package}
\name{funcharts-package}
\alias{funcharts}
\alias{funcharts-package}
\title{funcharts: Functional Control Charts}
\description{
Provides functional control charts for statistical process monitoring of functional data, using the methods of Capezza et al. (2020) \doi{10.1002/asmb.2507} and Centofanti et al. (2021) \doi{10.1080/00401706.2020.1753581}. The package is thoroughly illustrated in the paper of Capezza et al (2023) \doi{10.1080/00224065.2023.2219012}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/unina-sfere/funcharts}
  \item Report bugs at \url{https://github.com/unina-sfere/funcharts/issues}
}

}
\author{
\strong{Maintainer}: Christian Capezza \email{christian.capezza@unina.it}

Authors:
\itemize{
  \item Fabio Centofanti
  \item Antonio Lepore
  \item Alessandra Menafoglio
  \item Biagio Palumbo
  \item Simone Vantini
}

}
\keyword{internal}
