% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{nbasis}
\alias{nbasis}
\title{Number of basis functions}
\usage{
nbasis(object, ...)
}
\arguments{
\item{object}{An object from which to extract the number of basis functions.}

\item{...}{Further arguments passed to methods (not used).}
}
\description{
Generic function to extract the number of basis functions from an object.
}
