% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{nvar}
\alias{nvar}
\title{Number of variables}
\usage{
nvar(object, ...)
}
\arguments{
\item{object}{An object from which to extract the number of variables.}

\item{...}{Further arguments passed to methods (not used).}
}
\description{
Generic function to extract the number of variables from an object.
}
