% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mfd.R
\name{plot_mfd}
\alias{plot_mfd}
\title{Plot a Multivariate Functional Data Object.}
\usage{
plot_mfd(mfdobj)
}
\arguments{
\item{mfdobj}{A multivariate functional data object of class mfd.}
}
\value{
an object of class ggplot, created using
\code{ggplot() + geom_mfd(mfdobj = mfdobj)}.
}
\description{
Plot an object of class \code{mfd} using \code{ggplot2}.
}
\examples{
library(funcharts)
mfdobj <- data_sim_mfd()
plot_mfd(mfdobj)

}
\seealso{
\code{\link{geom_mfd}}
}
