% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_funcharts}
\alias{sim_funcharts}
\title{Simulate example data for funcharts}
\usage{
sim_funcharts()
}
\value{
A list with three objects, \code{datI} contains the phase I data,
\code{datI_tun} contains the tuning data,
\code{datII} contains the phase II data.
In the phase II data, the first group of 20 observations are in control,
the second group of 20 observations contains a moderate mean shift,
while the third group of 20 observations contains a severe mean shift.
The shift types are described in the paper from Centofanti et al. (2020),
}
\description{
Function used to simulate three data sets to illustrate the use of \code{funcharts}.
It uses the function \code{\link[funcharts]{simulate_mfd}},
which creates a data set with three functional covariates,
a functional response generated as a function of the three functional covariates,
and a scalar response generated as a function of the three functional covariates.
This function generates three data sets, one for phase I (1000 observations),
one for tuning (1000 observations), i.e.,
to estimate the control chart limits, and one for phase II monitoring (60 observations).
see also \code{\link[funcharts]{simulate_mfd}}.
}
\references{
Centofanti F, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2020)
Functional Regression Control Chart.
\emph{Technometrics}. <doi:10.1080/00401706.2020.1753581>
}
